Docker: Add docker support for building
This commit adds support for building webrcade-app-common in a docker
container.

There are 3 scripts that handle the bulk of the work here:

docker-run.sh: This is the user interactable script. It will build the
docker image, and run it. The output of this will be a
webrcade-app-common.tgz in a local "out/" folder

scripts/build-docker-image.sh: Build the docker image locally

scripts/build-webracde-app-common.sh: This script gets built in to the
docker image and handles actually building webrcade-app-common